Score system & logging BIG update
Many changes to score system and the way results are logged. System is not as nice as I would want, but idgaf - this was the most efficient way of achieving what I wanted. All points calculators now rely on an array of values, one value per game. Previously, they were incrementing a double, which would hamper our ability to do statistics (no info on variance, just a total and average). Fixed the last issues with logging personality values (added a title to the log files).